+++ scripts/build.js
@@ -56,7 +56,7 @@ const outlineExtensionsExternals = outlineExtensions.map((node) => {
 const outlineReactModules = fs
   .readdirSync(path.resolve('./packages/outline-react/src'))
   .map((str) => path.basename(str, '.js'))
-  .filter((str) => !str.includes('__tests__'));
+  .filter((str) => !str.includes('__tests__') && !str.includes('test-utils'));
 const outlineReactModuleExternals = outlineReactModules.map((module) => {
   const external = `outline-react/${module}`;
   wwwMappings[external] = module;
